Implement (but do not use yet) correction to low energy part of spectrum.
The correction is applied only to `JEOL_pts.spectrum()` and not to the
data cube itself. This makes sense as it would be time consuming and the
maximum difference is less than about 2 channels at the very lowest
energies. For maps this is irrelevant even when mapping a certain element.

NOTE: Still minor difference to `JEOL_pts.ref_spec` present. See comment
      in code.

Signed-off-by: Ivo Alxneit <ivo.alxneit@psi.ch>