Prevent SyntaxError when calculating 2D histogram with a single bin
Histograms with a single bin raise a SyntaxError:

Traceback (most recent call last):

  File "<ipython-input-55-e7da51c65b5f>", line 3, in <module>
    bins=bins)

  File "<__array_function__ internals>", line 6, in histogram2d

  File "/home/alxneit/share/miniconda3/lib/python3.7/site-packages/numpy/lib/twodim_base.py", line 713, in histogram2d
    hist, edges = histogramdd([x, y], bins, range, normed, weights, density)

  File "<__array_function__ internals>", line 6, in histogramdd

  File "/home/alxneit/share/miniconda3/lib/python3.7/site-packages/numpy/lib/histograms.py", line 1048, in histogramdd
    '`bins[{}]` must be positive, when an integer'.format(i))

ValueError: `bins[0]` must be positive, when an integer



If all shifts are zero
	bins = array([-0.5,  0.5])
	extent = [0, 0, 0, 0]

This bins-specifier is interpreted as [nx, ny] (number of bins) by
numpy.histogram2d. But integers are expected in this case! The all-zero
`extent` would only raise a UserWarning in matplotlib.pyplot.imshow().

If all shifts are identical but non-zero, `amax` is also non-zero and all
is fine.

To prevent this we set `amax` = 1 and then
	bins = array([-1.5, -0.5,  0.5,  1.5])
	extent = [-1, 1, -1, 1]

Signed-off-by: Ivo Alxneit <ivo.alxneit@psi.ch>